serial1/requirements.txt
alembic==1.4.3
aniso8601==8.0.0
antiorm==1.2.1
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
db==0.1.1
db-sqlite3==0.0.1
distlib==0.3.1
Django==3.1.3
django-filter==2.4.0
django-mysql==3.9.0
django-restframework==0.0.1
djangorestframework==3.12.2
dnspython==2.0.0
email-validator==1.1.1
Faker==4.1.3
filelock==3.0.12
flash==1.0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.9.0
Flask-DebugToolbar==0.11.0
Flask-Flash==1.6.8
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
httpie==2.3.0
idna==2.10
inflect==4.1.0
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
numpy==1.19.0
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.2
PyMySQL==0.10.1
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
virtualenv==20.0.26
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
WTForms==2.3.3
